requirements.txt
flask
torch
torchvision
pandas
Pillow
mediapipe 
opencv-python
numpy
scikit-learn
matplotlib